:hammer: refactor parser of push event
This commit is to refactor GitHubActionParser class which change partial
class for splitting many classes per trigger events.
