+++ src/Pact.Cli/Prsers/GitHubActionParser.cs
@@ -9,7 +9,7 @@ namespace Pact.Cli.Parsers;
 /// <summary>
 /// Parser class of GitHub Actions workflow file.
 /// </summary>
-public class GitHubActionParser
+public partial class GitHubActionParser
 {
     /// <summary>
     /// Initialize new instance of <see cref="GitHubActionParser" /> class.
@@ -165,65 +165,6 @@ public class GitHubActionParser
             else if (DynamicUtil.Is<string>(property, out string type))
             {
                 eventInfo.Types = new string[] { type };
-            }
-        }
-
-        return eventInfo;
-    }
-
-    private static PushTriggerEvent ParsePushEvent(dynamic yaml)
-    {
-        var eventInfo = new PushTriggerEvent();
-        if (yaml is null)
-        {
-            return eventInfo;
-        }
-
-        if (yaml.ContainsKey("branches"))
-        {
-            if (DynamicUtil.Is<List<object>>(yaml["branches"], out List<object> list))
-            {
-                eventInfo.Branches = list.Where(x => x is string).Select(x => (string)x).ToArray();
-            }
-        }
-
-        if (yaml.ContainsKey("tags"))
-        {
-            if (DynamicUtil.Is<List<object>>(yaml["tags"], out List<object> list))
-            {
-                eventInfo.Tags = list.Where(x => x is string).Select(x => (string)x).ToArray();
-            }
-        }
-
-        if (yaml.ContainsKey("paths"))
-        {
-            if (DynamicUtil.Is<List<object>>(yaml["paths"], out List<object> list))
-            {
-                eventInfo.Paths = list.Where(x => x is string).Select(x => (string)x).ToArray();
-            }
-        }
-
-        if (yaml.ContainsKey("branches-ignore"))
-        {
-            if (DynamicUtil.Is<List<object>>(yaml["branches-ignore"], out List<object> list))
-            {
-                eventInfo.BranchesIgnore = list.Where(x => x is string).Select(x => (string)x).ToArray();
-            }
-        }
-
-        if (yaml.ContainsKey("tags-ignore"))
-        {
-            if (DynamicUtil.Is<List<object>>(yaml["tags-ignore"], out List<object> list))
-            {
-                eventInfo.TagsIgnore = list.Where(x => x is string).Select(x => (string)x).ToArray();
-            }
-        }
-
-        if (yaml.ContainsKey("paths-ignore"))
-        {
-            if (DynamicUtil.Is<List<object>>(yaml["paths-ignore"], out List<object> list))
-            {
-                eventInfo.PathsIgnore = list.Where(x => x is string).Select(x => (string)x).ToArray();
             }
         }
